Greatly improve the usability of the completion
Don't change the completion configuration wily-nilly between
invocations. This allows the completion function to be used in a sane
way, as a regular omni-function, and not require the usage of arcane
file-specific mappings.
Change the default completion type to 'syntax'. The current default,
'table' blurts an error and SLEEPS for 2 seconds without that other
obscure perl extension that doesn't come by default anywhere.

This should be coupled with another setting:
- On the vimrc, `let g:omni_sql_no_default_maps = 1` to stop the
  creation of those weird mappings